+++ src/cbc_sdk/__init__.py
@@ -7,5 +7,4 @@ __copyright__ = 'Copyright 2020-2021 VMware Carbon Black'
 __version__ = '1.0.1'
 
 from .rest_api import CBCloudAPI
-from .base import UnrefreshableModel
 from .cache import lru
