+++ src/cbc_sdk/enterprise_edr/__init__.py
@@ -1,8 +1,5 @@
 from __future__ import absolute_import
 
-from cbc_sdk.enterprise_edr.base import Query
-
-
 
 from cbc_sdk.enterprise_edr.threat_intelligence import (Watchlist, Feed, Report,
                                                       ReportSeverity, IOC, IOC_V2,
