+++ src/cbc_sdk/platform/__init__.py
@@ -1,8 +1,13 @@
 from __future__ import absolute_import
 
-from cbc_sdk.platform.base import (PlatformModel, PlatformQueryBase, Process,
-                                   Event, ProcessFacet, AsyncProcessQuery,
-                                   SummaryQuery, EventFacet)
-from cbc_sdk.platform.devices import Device, DeviceSearchQuery
+from cbc_sdk.platform.base import PlatformModel
+
 from cbc_sdk.platform.alerts import (BaseAlert, WatchlistAlert, CBAnalyticsAlert,
-                                     VMwareAlert, Workflow, WorkflowStatus)
+                                     Workflow, WorkflowStatus)
+
+from cbc_sdk.platform.devices import Device, DeviceSearchQuery
+
+from cbc_sdk.platform.events import Event, EventFacet
+
+from cbc_sdk.platform.processes import (Process, ProcessFacet,
+                                        AsyncProcessQuery, SummaryQuery)
