+++ src/tests/unit/platform/test_alertsv6_api.py
@@ -13,7 +13,7 @@
 
 import pytest
 from cbc_sdk.errors import ApiError
-from cbc_sdk.platform import BaseAlert, CBAnalyticsAlert, VMwareAlert, WatchlistAlert, WorkflowStatus
+from cbc_sdk.platform import BaseAlert, CBAnalyticsAlert, WatchlistAlert, WorkflowStatus
 from cbc_sdk.rest_api import CBCloudAPI
 from tests.unit.fixtures.stubresponse import StubResponse, patch_cbc_sdk_api
 
@@ -271,78 +271,6 @@ def test_query_cbanalyticsalert_invalid_criteria_values(method, arg):
         meth(arg)
 
 
-def test_query_vmwarealert_with_all_bells_and_whistles(monkeypatch):
-    """Test a VMware alert query with all options selected."""
-    _was_called = False
-
-    def _run_query(url, body, **kwargs):
-        nonlocal _was_called
-        assert url == "/appservices/v6/orgs/Z100/alerts/vmware/_search"
-        assert body == {"query": "Blort",
-                        "rows": 100,
-                        "criteria": {"category": ["SERIOUS", "CRITICAL"], "device_id": [6023], "device_name": ["HAL"],
-                                     "device_os": ["LINUX"], "device_os_version": ["0.1.2"],
-                                     "device_username": ["JRN"], "group_results": True, "id": ["S0L0"],
-                                     "legacy_alert_id": ["S0L0_1"], "minimum_severity": 6, "policy_id": [8675309],
-                                     "policy_name": ["Strict"], "process_name": ["IEXPLORE.EXE"],
-                                     "process_sha256": ["0123456789ABCDEF0123456789ABCDEF"],
-                                     "reputation": ["SUSPECT_MALWARE"], "tag": ["Frood"], "target_value": ["HIGH"],
-                                     "threat_id": ["B0RG"], "type": ["WATCHLIST"], "workflow": ["OPEN"],
-                                     "group_id": [14]}, "sort": [{"field": "name", "order": "DESC"}]}
-        _was_called = True
-        return StubResponse({"results": [{"id": "S0L0", "org_key": "Z100", "threat_id": "B0RG",
-                                          "workflow": {"state": "OPEN"}}], "num_found": 1})
-
-    api = CBCloudAPI(url="https://example.com", token="ABCD/1234", org_key="Z100", ssl_verify=True)
-    patch_cbc_sdk_api(monkeypatch, api, POST=_run_query)
-    query = api.select(VMwareAlert).where("Blort").set_categories(["SERIOUS", "CRITICAL"]).set_device_ids([6023]) \
-        .set_device_names(["HAL"]).set_device_os(["LINUX"]).set_device_os_versions(["0.1.2"]) \
-        .set_device_username(["JRN"]).set_group_results(True).set_alert_ids(["S0L0"]) \
-        .set_legacy_alert_ids(["S0L0_1"]).set_minimum_severity(6).set_policy_ids([8675309]) \
-        .set_policy_names(["Strict"]).set_process_names(["IEXPLORE.EXE"]) \
-        .set_process_sha256(["0123456789ABCDEF0123456789ABCDEF"]).set_reputations(["SUSPECT_MALWARE"]) \
-        .set_tags(["Frood"]).set_target_priorities(["HIGH"]).set_threat_ids(["B0RG"]).set_types(["WATCHLIST"]) \
-        .set_workflows(["OPEN"]).set_group_ids([14]).sort_by("name", "DESC")
-    a = query.one()
-    assert _was_called
-    assert a.id == "S0L0"
-    assert a.org_key == "Z100"
-    assert a.threat_id == "B0RG"
-    assert a.workflow_.state == "OPEN"
-
-
-def test_query_vmwarealert_facets(monkeypatch):
-    """Test a VMware alert facet query."""
-    _was_called = False
-
-    def _run_facet_query(url, body, **kwargs):
-        nonlocal _was_called
-        assert url == "/appservices/v6/orgs/Z100/alerts/vmware/_facet"
-        assert body == {"query": "Blort", "criteria": {"workflow": ["OPEN"]},
-                        "terms": {"rows": 0, "fields": ["REPUTATION", "STATUS"]},
-                        "rows": 100}
-        _was_called = True
-        return StubResponse({"results": [{"field": {},
-                                          "values": [{"id": "reputation", "name": "reputationX", "total": 4}]},
-                                         {"field": {},
-                                          "values": [{"id": "status", "name": "statusX", "total": 9}]}]})
-
-    api = CBCloudAPI(url="https://example.com", token="ABCD/1234", org_key="Z100", ssl_verify=True)
-    patch_cbc_sdk_api(monkeypatch, api, POST=_run_facet_query)
-    query = api.select(VMwareAlert).where("Blort").set_workflows(["OPEN"])
-    f = query.facets(["REPUTATION", "STATUS"])
-    assert _was_called
-    assert f == [{"field": {}, "values": [{"id": "reputation", "name": "reputationX", "total": 4}]},
-                 {"field": {}, "values": [{"id": "status", "name": "statusX", "total": 9}]}]
-
-
-def test_query_vmwarealert_invalid_group_ids():
-    """Test error message for invalid group IDs for VMware alert queries."""
-    api = CBCloudAPI(url="https://example.com", token="ABCD/1234", org_key="Z100", ssl_verify=True)
-    with pytest.raises(ApiError):
-        api.select(VMwareAlert).set_group_ids(["Bogus"])
-
-
 def test_query_watchlistalert_with_all_bells_and_whistles(monkeypatch):
     """Test a watchlist alert query with all options selected."""
     _was_called = False
@@ -504,7 +432,7 @@ def test_alerts_bulk_dismiss_vmware(monkeypatch):
 
     def _do_dismiss(url, body, **kwargs):
         nonlocal _was_called
-        assert url == "/appservices/v6/orgs/Z100/alerts/vmware/workflow/_criteria"
+        assert url == "/appservices/v6/orgs/Z100/alerts/cbanalytics/workflow/_criteria"
         assert body == {"query": "Blort", "state": "DISMISSED", "remediation_state": "Fixed", "comment": "Yessir",
                         "criteria": {"device_name": ["HAL9000"]}}
         _was_called = True
@@ -512,7 +440,7 @@ def test_alerts_bulk_dismiss_vmware(monkeypatch):
 
     api = CBCloudAPI(url="https://example.com", token="ABCD/1234", org_key="Z100", ssl_verify=True)
     patch_cbc_sdk_api(monkeypatch, api, POST=_do_dismiss)
-    q = api.select(VMwareAlert).where("Blort").set_device_names(["HAL9000"])
+    q = api.select(CBAnalyticsAlert).where("Blort").set_device_names(["HAL9000"])
     reqid = q.dismiss("Fixed", "Yessir")
     assert _was_called
     assert reqid == "497ABX"
